+++ src/CoState.cs
@@ -104,7 +104,7 @@ namespace FSM
 			onExit?.Invoke(this);
 		}
 
-		public override void RequestExit()
+		public override void OnExitRequest()
 		{
 			if (!needsExitTime || (canExit != null && canExit(this)))
 			{